backend/requirements.txt
fastapi
uvicorn
python-jose[cryptography]
bcrypt
python-dotenv
supabase